Enable SystemUIGoogleScreenshotTests on main and disable on udc-qpr-dev
At the moment the tests are only running on `udc-qpr-dev`, and not
on `main`.
If we disable them on `udc-qpr-dev` and keep them enabled on `main`,
they will still run in presubmits of both branches.

This way we also solve the problem of having potentially different
golden differences in both branches.

Change-Id: Ic7c389084838e57008d7eb327e674ef6407975af
Fixes: 292206289
Merged-In: I86236180877f14273f2b59fc3d4b37529ed87c49