Add new FeatureFlagsClassic alias
Deprecate the old FeatureFlags name.

This change is in preparation for the introduction of the new
aconfig-based flagging system, which also utilizes a class called
FeatureFlags. To prevent the need for qualified names, as well as
general naming confusion, we are moving usage of the "classic" feature
flags to a new name going forward.

Bug: 292511372
Test: make
Change-Id: Ieeb339c0311cd7a81aacd68e8808ee812b6bb018