Filter AppMetadata file from Session.getNames
AppMetadata for a install is set using setAppMetadata and the bundle is
written to a file. This file should not be returned included in the
getNames API since it was not added to the session using openWrite.

Bug: 296471104
Test: atest InstallAppMetadataTest
Merged-In: Ia089884ac91b008e4179bd9970eb6c8c91daaf25
Change-Id: Ia089884ac91b008e4179bd9970eb6c8c91daaf25
(cherry picked from commit 0bd19d31e38bb2c9b8bcce152ad123170371ace4)