Delete WrappedSysUiStatsEvent in favor of new StatsEventTestUtils
The custom WrappedSysUiStatsEvent was used to enable testing of pulled atoms, which otherwise were only available in a previously-uninspectable StatsEvent format. Now that there is an official test utility for converting StatsEvent objects to the associated Atom (which we can read), we can inspect the contents of generated StatsEvent directly, removing the need for wrapping the builder and maintaining separate custom testing code.

Also suppresses lint warning on ZenModeHelperTest so we don't have to lock the locks to set up mConfig.

Bug: 161127189
Test: atest ZenModeHelperTest, PreferencesHelperTest

Change-Id: Ibb647adc635fcebf628c12aa0c47392a7cdc3424