Support key gesture detection for stylus tail button in SysUI
Here, we add gesture detection logic for the stylus tail button in
SysUI.

The notes task will only be shown immediately on the UP event of the
first key press of a key gesture. A long press will not do anything for
now. If there is a multi-press (e.g. double-press, triple-press), only
the first press will react to the event.

Bug: 298056902
Test: manual with stylus
Test: atest NoteTaskInitializerTest
Change-Id: I90391f411d84ec5c749bb4d8cdefd7e610539e6e