fix(status bar): wifi and mobile icons flickering in large font/display size
If the status bar status icons container contains only wifi and mobile views, and the container max width is only enough for the mobile view, the expected ui result is the wifi view becomes hidden and the mobile view shows the dot. However, setting the wifi view isVisibile to false causes its width becomes 0, then the container would mis-consider the max width is enough and try to reshow both views. After the container mis-shows both views, it would find the max width is not enough and try to hide the wifi view and make the mobile view show the dot. Then the loop happens and causes the flickering.

Therefore, for a quick workaround, instead of setting the wifi view isVisible to false, we set its visibility to INVISIBLE to achieve the same behavior and keep its width, so the container would do the correct calculation and eliminate the flicker.

Bug: 296864006
Test: manually - attached video in bug
      atest ModernStatusBarWifiViewTest
      atest ModernStatusBarMobileViewTest
Change-Id: Id6e00facda2c4aed470f772741c1d492eebad17d
Merged-In: Id6e00facda2c4aed470f772741c1d492eebad17d