Block relayout during Recents transition
The Recents transition takes ownership of the task leash to reposition
it and scale it if needed for the overview animation. During this time,
a focus change may trigger TaskListener#onTaskInfoChanged which causes
a window decor relayout that applies an incorrect position and crop.

Bug: 296921174
Test: swipe up from taskbar, verify freeform task doesn't jump out of
position and gets a bad crop applied.
Test: atest DesktopModeWindowDecorViewModel

Change-Id: If6af9ab76e9c7174a9bb8c9c97e4a79a9b52a775