Add ServiceThread to WallpaperManagerService.
During a user switch, binding the WallpaperService was done on
the system_server's main handler, and was being delayed unnecessarily
when the system_server was busy. With this CL, we are giving the
WallpaperManagerService its own thread handler and do the binding of
the target user's wallpaper service on that handler. This improves
user switch duration by 17%, and reduces spikiness 10 times.

Bug: 300421828
Test: Manual control of Perfetto traces
Change-Id: Ib8514d3cdc7a69359866b1e3d0c8ffcb32c9a4aa