Support bypassing interruption policy for scroll haptic feedback
The scroll haptic feedback should be played regardless of interruption
mode (as they are a system navigation feedback). This change first makes
changes to the VibratorManagerService to enable haptic feedback
vibrations to use bypass flags without any permission, and then enables
interruption policy bypassing for the SCROLL_* constants.

The guard on SCROLL_* constants' bypassing of interruption policy uses
the same flag as the one used for the scroll feedback APIs.

Bug: 289480045
Test: atest HapticFeedbackVibrationProviderTest
Test: atest VibratorManagerServiceTest
Change-Id: I1eb6368c3685e71eead334644d973a9d21f6d207