Adjust keyguard preview renderer to allow different displays
Make adjustments so keyguard preview renderer can create a preview for
any specified display. This enables previewing the lock screen workspace
for multi-display devices, needed in wallpaper picker's new preview UI.

We enable this by making sure views are created using
a context that is adjusted for the display, instead of the current
context.

Bug: 291761856
Bug: 303313708
Test: manually verified in new wallpaper picker preview UI
Change-Id: I786f93259191666550158a301e1bcc35a1ae7af3