Add module lib and test union java_api_library module
Currently, hiddenapi_modular defines test api surface as a subset of
module lib api surface, causing hiddenapi check failure in from-text
stub build. In order to mitigate this and enable the hiddenapi check in
from-text stub build, this change introduces a java_api_library module
that generates the stub jar that contains both test and module lib
api contributions. The module will not be available for consumption, but
will only be used in hiddenapi check.

Test: m android_test_module_lib_stubs_current.from-text --build-from-text-stub
Bug: 191644675
Change-Id: I2b6c9f75d9ebcfd75fa5b509dbef14aa0f88ac97