Clear calling identity for trusted change from keyguard
It is an internal method (not aidl of WMS) which can only be called
from keyguard service (IKeyguardStateCallback). And it is not
sensitive to refresh visibility based on the current state of system.

Otherwise if there is an activity that allows to resume after the
state change, it could causes random SecurityException if systemui
doesn't declare to use such system permission.

Bug: 302961176
Test: Simulate triggering IKeyguardStateCallback#onTrustChanged
      from keyguard.
Change-Id: I0a98937d0c02aa0662ec168c9088335738fe779d