Move updateAudioProfile work to bg
updateAudioProfile makes calls to determine state of
CachedBluetoothDevice. Those calls need to happe in the background. As
this is called from the callback in updateConnected (which ends up being
called in main thread), we need to send this off thread.

Also, add annotations on methods we know should be called from a bg
thread.

Test: atest BluetoothControllerImplTest
Test: perfetto trace (no calls in main thread after continuation)
Fixes: 304103830
Fixes: 289430339
Change-Id: Ie1cdbdf63ea6731c553c827fcdbfb8bef340c400